requirements.txt
python-telegram-bot==13.0
boto3==1.16.23